requirements.txt
breadability==0.1.20
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
discord-webhook==1.3.0
docopt==0.6.2
fonttools==4.53.1
gensim==4.3.3
idna==3.7
jira==3.8.0
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
plotly==5.22.0
pycountry==24.6.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
schedule==1.2.2
scipy==1.13.1
six==1.16.0
smart-open==7.0.4
sumy==0.11.0
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
